Create data file when needed in List

+++ lib/tracking/list.rb
@@ -18,9 +18,13 @@ module Tracking
 		#
 		# @return [Array] an array of arrays (tasks), each consisting of two strings (the tasks's start time and the task's name)
 		def get
-			tasks = CSV.read(@data_file, @csv_options)
-			tasks = tasks[-Config[:lines]..-1] if tasks.length > Config[:lines]
-			return tasks
+			if File.exist? @data_file
+				tasks = CSV.read(@data_file, @csv_options)
+				tasks = tasks[-Config[:lines]..-1] if tasks.length > Config[:lines]
+				return tasks
+			else
+				return []
+			end
 		end
 
 		# Adds a task to the list
@@ -28,6 +32,7 @@ module Tracking
 		# @param [String] task the name of the task to add to the list
 		def add task
 			time = Time.now.to_s
+			FileUtils.touch @data_file unless File.exist? @data_file
 			File.open(@data_file, 'a') do |file|
 				file << [ time, task ].to_csv(@csv_options)
 			end
@@ -35,19 +40,23 @@ module Tracking
 
 		# Deletes the last task from the list
 		def delete
-			lines = File.readlines @data_file
-			lines.pop # Or delete specific lines in the future
-			File.open(@data_file, 'w') do |file| 
-				lines.each do |line|
-					file << line
+			if File.exist? @data_file
+				lines = File.readlines @data_file
+				lines.pop # Or delete specific lines in the future
+				File.open(@data_file, 'w') do |file| 
+					lines.each do |line|
+						file << line
+					end
 				end
 			end
 		end
 
 		# Clears the entire list
 		def clear
-			FileUtils.rm @data_file
-			FileUtils.touch @data_file
+			if File.exist? @data_file
+				FileUtils.rm @data_file
+				FileUtils.touch @data_file
+			end
 		end
 
 		# Gets the elapsed time between two times and formats it into a string for
